Change container text sizes, colors, and spacing
The problem requires two text fields and a button to be on the
left side and on top of each other, with a placeholder for an
image on the right side. We accomplish this using flex layout,
with one flex item containing the text fields and button and the
other containing the placeholder for images.