Add spacing and content to abilities section
The old version had the basic template content from the given image.
It also had no spacing and structure. We added Yasuo's abilities,
excluding his passive, with their pictures and descriptions. We also
added spacing and structure similar to the templates.

Some content in the html was changed as we needed to add in extra divs
to help with structuring. For example, we were unable to put "Abilities"
on top of the abilities as there wasn't a div containing all the
abilities. So we added a div with class "boxes" to contain all the divs
with class "box" (aka abilities).